+++ lib/value_semantics/monkey_patched.rb
@@ -1,0 +1,3 @@
+require 'value_semantics'
+
+ValueSemantics.monkey_patch!