+++ CHANGELOG.md
@@ -5,7 +5,7 @@ Notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## Unreleased
+## [3.5.0] - 2020-08-17
 ### Added
 - Square bracket attr reader like `person[:name]`
 - `HashOf` built-in validator, similar to `ArrayOf`